[RestResourceHost] Extract incident related codes to RestResourceIncident
