+++ server/src/FaceRest.cc
@@ -42,6 +42,7 @@
 #include "RestResourceSystem.h"
 #include "RestResourceAction.h"
 #include "RestResourceHost.h"
+#include "RestResourceIncident.h"
 #include "RestResourceIncidentTracker.h"
 #include "RestResourceServer.h"
 #include "RestResourceSeverityRank.h"
@@ -402,6 +403,7 @@ gpointer FaceRest::mainThread(HatoholThreadArg *arg)
 	RestResourceUser::registerFactories(this);
 	RestResourceServer::registerFactories(this);
 	RestResourceHost::registerFactories(this);
+	RestResourceIncident::registerFactories(this);
 	RestResourceAction::registerFactories(this);
 	RestResourceIncidentTracker::registerFactories(this);
 	RestResourceSeverityRank::registerFactories(this);
